requirements.txt
opencv-python
mediapipe
numpy
torch
torchvision
Pillow